==> deu/11_deu_explore.Rmd <==
---
title: "11_deu_explore"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(scales)
library(ggunchained)
library(lmerTest)
library(cowplot)
library(rethinking)
```

FILE PATHS
```{r}
ntr_paths = list.files("deu/data/ntr_out", pattern = "*.csv", full.names = T)
shuff_ntr_paths = list.files("deu/data/ntr_shuffled/", pattern = "*.csv", full.names = T)
```


PARTICULAR DATASET W41
```{r}
norm41 = read_csv(ntr_paths[str_detect(ntr_paths, "41")])
shuff41 = read_csv(shuff_ntr_paths[str_detect(shuff_ntr_paths, "41")])
```


DISTRIBUTIONS
```{r}
distribution_plots <- function(w, iso) {
  
  ### filepaths
  ord_subpath = list.files(paste0(iso, "/data/ntr_out"), 
                           pattern = paste0("W", w, "_"), full.names = T)
  
  shu_subpath = list.files(paste0(iso, "/data/ntr_shuffled"), 
                           pattern = paste0("W", w, "_"), full.names = T)
  
  norm = read_csv(ord_subpath)
  shuff = read_csv(shu_subpath)
  
  
  ### MAKE PLOTS
  # ORDERED
  ord_N = ggplot(norm, aes(z_novelty)) +
    geom_density() +
    theme_classic() +
    labs(title = "Ordered Novelty")
  
  ord_T = ggplot(norm, aes(z_transience)) +
    geom_density() +
    theme_classic() +
    labs(title = "Ordered Transience")
  
  ord_R = ggplot(norm, aes(z_resonance)) +
    geom_density() +
    theme_classic() +
    labs(title = "Ordered Resonance")
  
  # SHUFFLED
  shu_N = ggplot(shuff, aes(z_novelty)) +
    geom_density() +
    theme_classic() +
    labs(title = "Shuffled Novelty")
  
  shu_T = ggplot(shuff, aes(z_transience)) +
    geom_density() +
    theme_classic() +
    labs(title = "Shuffled Transience")
  
  shu_R = ggplot(shuff, aes(z_resonance)) +
    geom_density() +
    theme_classic() +
    labs(title = "Shuffled Resonance")
  
  
  ### ARRANGE PLOTS INTO GRID
  plot_row = plot_grid(ord_N, shu_N, ord_T, shu_T, ord_R, shu_R,
                       nrow = 3,
                       ncol = 2)
  
  title <- ggdraw() + 
    draw_label(
      paste0("[", toupper(iso), "] ", "w = ", w),
      fontface = 'bold',
      x = 0,
      hjust = 0
    ) +
    theme(
      # add margin on the left of the drawing canvas,
      # so title is aligned with left edge of first plot
      plot.margin = margin(0, 0, 0, 7)
    )
  
  final = plot_grid(
    title, plot_row,
    ncol = 1,
    # rel_heights values control vertical title margins
    rel_heights = c(0.1, 1)
  )
  
  cowplot::ggsave2(plot = final, 
                   filename = paste0(iso, "/data/plots/dens/", w, "_dens.png"))
  
}

#w_range = as.character(seq(1, 100, 5))
w_range = c(300, 900)

lapply(w_range,
       distribution_plots,
       iso = "deu")
```


PATTERNS
```{r}
export_plots <- function(path) {
  
  # load in
  ntr_w = read_csv(path)
  
  # extract timeframe
  w = str_extract(path, "(?<=W)\\d+")
  
  
  # RESONANCE ~ NOVELTY
  ggplot(ntr_w, aes(z_novelty, z_resonance)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000),
                         labels = trans_format("log10", 
                                               math_format(expr = 10^.x, format = force))) +
    geom_smooth(method = "lm", colour = "black", alpha = 0, size = 0.5) +
    labs(x = "Novelty (z-scaled)\n", y = "Resonance (z-scaled)", 
         title = paste0("Resonance vs. Novelty, w = ", w),
         subtitle = "with a regression line",
         caption = "Data source: Bundestag debates (1998-2019)",
         fill = "speech\ncount") +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 6, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )
  
  ggsave(filename = paste0("rn_", w, ".png"), path = "deu/data/plots/RN/")
  
  
  
  # NOVELTY VS TRANSIENCE
  ggplot(ntr_w, aes(z_novelty, z_transience)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                         labels = trans_format("log10", math_format(expr = 10^.x, format = force))) +
    geom_abline(colour = "black", linetype = "dashed") +
    labs(x = "Novelty (z-scaled)\n", y = "Transience (z-scaled)", 
         title = paste0("Novelty vs. Transience, w =", w),
         subtitle = "with an identity line (x = y)",
         caption = "Data source: Bundestag debates (1998-2019)",
         fill = "speech\ncount") +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 8, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )
  
  ggsave(filename = paste0("nt_", w, ".png"), path = "deu/data/plots/NT/")
  
  
}

export_shuff_plots <- function(path) {
  
  # load in
  ntr_w = read_csv(path)
  
  # extract timeframe
  w = str_extract(path, "(?<=W)\\d+")
  
  
  # RESONANCE ~ NOVELTY
  ggplot(ntr_w, aes(z_novelty, z_resonance)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000),
                         labels = trans_format("log10", 
                                               math_format(expr = 10^.x, format = force))) +
    geom_smooth(method = "lm", colour = "black", alpha = 0, size = 0.5) +
    labs(x = "Novelty (z-scaled)\n", y = "Resonance (z-scaled)", 
         title = paste0("Shuffled Resonance vs. Novelty, w = ", w),
         subtitle = "with a regression line",
         caption = "Data source: Bundestag debates (1998-2019)",
         fill = "speech\ncount") +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 6, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )
  
  ggsave(filename = paste0("shuff_rn_", w, ".png"), path = "deu/data/plots/shuffled/")
  
  
  
  # NOVELTY VS TRANSIENCE
  ggplot(ntr_w, aes(z_novelty, z_transience)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                         labels = trans_format("log10", math_format(expr = 10^.x, format = force))) +
    geom_abline(colour = "black", linetype = "dashed") +
    labs(x = "Novelty (z-scaled)\n", y = "Transience (z-scaled)", 
         title = paste0("Shuffled Novelty vs. Transience, w =", w),
         subtitle = "with an identity line (x = y)",
         caption = "Data source: Bundestag debates (1998-2019)",
         fill = "speech\ncount") +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 8, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )
  
  ggsave(filename = paste0("shuff_nt_", w, ".png"), path = "deu/data/plots/shuffled/")
  
  
}

lapply(ntr_paths, export_plots)
lapply(shuff_ntr_paths, export_shuff_plots)
```


# delta R
```{r}
D_41 %>%
  ggplot(aes(z_novelty, delta_R)) +
   geom_bin2d(bins = 100) +
  scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                labels = trans_format("log10", math_format(expr = 10^.x, format = force)))

D_41S %>%
  ggplot(aes(z_novelty, delta_R)) +
  geom_bin2d(bins = 100) +
  scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                labels = trans_format("log10", math_format(expr = 10^.x, format = force)))
```

# TESTS
## DEU vs. FRA
```{r}
D_41 = D_41 %>%
  mutate(nation = "Germany")

F_41 = read_csv("fra/data/ntr_out/W41__from_matrix.csv") %>%
  mutate(nation = "France")

FD = bind_rows(D_41, F_41)

big = lmer(z_resonance ~ 1 + nation + (1 + novelty | nation), 
           data = FD,
           REML = FALSE)

summary(big)

```

## SHUFF vs. ORD
```{r}
de_ord = lmer(z_resonance ~ z_novelty + (1 | name), data = D_41, REML = FALSE)
de_shuff = lmer(z_resonance ~ z_novelty + (1 | name), data = D_41S, REML = FALSE)

summary(de_ord)
summary(de_shuff)

confint(de_ord)
confint(de_shuff)

BIC(de_ord)
BIC(de_shuff)
```

```{r}
F_41S = read_csv("fra/data/ntr_shuffled/W41__ntr_shuf.csv")

fr_ord = lmer(z_resonance ~ z_novelty + (1 | name), data = F_41, REML = FALSE)
fr_shuff = lmer(z_resonance ~ z_novelty + (1 | name), data = F_41S, REML = FALSE)

summary(fr_ord)
summary(fr_shuff)

confint(fr_ord)
confint(fr_shuff)

BIC(fr_ord)
BIC(fr_shuff)
```
